adicion de about us en el modal
